Fully specify
