+++ src/DotNetProjectFile.Analyzers/DotNetProjectFile.Analyzers.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <Import Project="../../props/common.props" />
 
@@ -68,7 +68,7 @@ ToBeReleased:
 - Report issues on root node only on start element.
 - Proj0036: Remove None when redundant. (NEW RULE)
 - Proj0037: Exclude runtime when all assets are private. (NEW RULE)
-- Proj0038: Specify NoWarn rule IDs fully. (NEW RULE)
+- Proj0038: Fully specify NoWarn rule IDs. (NEW RULE)
 - Proj0214: Severity to INFO as it is considered too opinionated. (UPDATE)
 - Proj0503: Package license is unknown. (NEW RULE)
 - Proj0504: Package license has changed. (NEW RULE)
